+++ src/main/java/ru/vsu/kinoprokat/rental/CreateRentalDto.java
@@ -1,0 +1,7 @@
+package ru.vsu.kinoprokat.rental;
+
+public record CreateRentalDto(
+        Long userId,
+        Long movieId
+) {
+}