+++ src/main/java/ru/vsu/kinoprokat/rental/RentalController.java
@@ -1,0 +1,20 @@
+package ru.vsu.kinoprokat.rental;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/rentals")
+@RequiredArgsConstructor
+public class RentalController {
+
+    private final RentalService rentalService;
+
+    @PostMapping
+    public RentalDto create(@RequestBody CreateRentalDto dto) {
+        return rentalService.create(dto);
+    }
+}