+++ src/main/java/ru/vsu/kinoprokat/rental/RentalRepository.java
@@ -1,0 +1,6 @@
+package ru.vsu.kinoprokat.rental;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RentalRepository extends JpaRepository<RentalEntity, Long> {
+}