+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2026-06-06
 ### Added
 - [movie-management] kinoprokat: добавить сущность фильм 
-- [movie-management] kinoprokat: добавить сущность пользователь
+- [user-management] kinoprokat: добавить сущность пользователь
+- [rental-management] kinoprokat: добавить функциональность проката фильма
 ### Fixed
 - [movie-create-dto] kinoprokat: поправить MovieCreateDto